ocr_service/requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
easyocr == 1.7.2
python-multipart == 0.0.20
pillow == 11.2.1
pytest == 8.3.5
httpx  == 0.28.1
